Update README.md
Imported from notes typed within the descriptor/alt-text field of an iCloud-hosted JPEG within an iOS 26.2 (23C55) version of the Photos app.  IYI The Author (Satoshi Nakamoto) has been using iOS since the 1st Generation iPhone, and sometimes it helps to simply start typing notes in whichever app is currently quoteunquote open instead of switching to another app and delaying the simple and altogether nice-in-a-Just-'Cause-way of jotting thoughts down on something of note; in this case justsohappenstancially the same JPEG (which was saved from  was used as wallpaper in a screenshots-of-lockscreen-variations-arranged-as-classical-triptych-only-a-la-iOS-Opera-browser-default-Fast-Action-Button-visualization(s).